This commit includes a variety of changes that improve the codebase's
quality and consistency. The following changes were made:

- Replaced `==` with `===` for strict comparisons.
- Removed redundant `break` statements after `return` in switch-case
structures.
- Fixed variable naming consistency and unnecessary variable
assignments.
- Simplified conditional statements and removed unused code.
- Corrected minor typos and formatted code for better readability.
- Enhanced type safety by casting variables explicitly and adding type
hints.
- Improved function and variable documentation with consistent comments.
- Refactored repeated logic to improve maintainability.
- Removed redundant checks for null values and unnecessary extractions.

+++ src/PersonEditor.php
@@ -23,7 +23,7 @@ $sPageTitle = gettext('Person Editor');
 // Get the PersonID out of the querystring
 $iPersonID = 0;
 if (array_key_exists('PersonID', $_GET)) {
-    $iPersonID = InputUtils::legacyFilterInput($_GET['PersonID'], 'int');
+    $iPersonID = (int) InputUtils::legacyFilterInput($_GET['PersonID'], 'int');
 }
 $isNewPerson = $iPersonID === 0;
